refactor(table-of-contents): extract heading-registry and active-heading state machines
Splits table-of-contents.svelte (643 lines) per #1179: the DOM-heading
derivation + MutationObserver retry state machine moves to
TableOfContentsHeadingRegistry, and the IntersectionObserver
active-heading scroll-spy moves to TableOfContentsActiveHeadingTracker.
Both are plain classes instantiated at component top level; the
component keeps two thin $effects that call each controller's sync().

These state machines were previously only reachable through a full
component mount plus real DOM/observer events. Adds
table-of-contents-heading-registry.test.ts and
table-of-contents-active-heading.test.ts covering the slugify/dedupe
and pickActiveId logic directly.

Pure code motion; behavior and public API unchanged.
